+++ FirebotProxy.Infrastructure/Adapters/RemoveMessagesByUsernameCommandHandler.cs
@@ -1,8 +1,11 @@
-﻿using FirebotProxy.Data.Access;
+﻿using System.Runtime.CompilerServices;
+using FirebotProxy.Data.Access;
 using FirebotProxy.SecondaryPorts.RemoveMessagesByUsername;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
+
+[assembly: InternalsVisibleTo("FirebotProxy.Infrastructure.Tests")]
 
 namespace FirebotProxy.Infrastructure.Adapters;
 
@@ -27,9 +30,11 @@ internal class RemoveMessagesByUsernameCommandHandler : IRequestHandler<RemoveMe
 
             try
             {
-                const string query = "DELETE FROM ChatMessages WHERE SenderUsername = {0}";
-
-                await _context.Database.ExecuteSqlRawAsync(query, request.SenderUsername);
+                if (_context.Database.IsRelational())
+                {
+                    const string query = "DELETE FROM ChatMessages WHERE SenderUsername = {0}";
+                    await _context.Database.ExecuteSqlRawAsync(query, request.SenderUsername);
+                }
 
                 await _context.SaveChangesAsync(cancellationToken);
